Move Vax plugin from anal to arch
